RLLib Training + Evaluation with wandb logging
Working example of the multiagent example extended with the evaluation of test_multiagent.
We use RLLib with Ray Tune to automatically track metrics and model checkpoints
Then we use the W&B API with rich media types to construct wandb.Tables containing the various recordings of the training
We follow up with logging of evaluation media for the best model on a test environment

Made changes to the environments and loggers to accommodate for better routing of constructed files and media

NOTE: Different environments work but multi drone evaluation errs!